fixed minor typos.

+++ _bootstrap/setup-bash.sh
@@ -93,11 +93,11 @@ cd $dir
 msg "Moving existing dotfiles from $HOME to $olddir..."
 for file in $files; do
   [ -f $HOME/.$file ] && mv $HOME/.$file $olddir
-  ln -s $dir/$file $HOME/.$file
+  ln -s $dir/$file $HOME/.${file}
 done
 
 for file in $nodot; do
-  mv $HOME/$file $olddir
+  [ -f $HOME/.$file ] && mv $HOME/$file $olddir
   ln -s $dir/$file $HOME/$file
   chflags -h hidden $HOME/$file
 done
